Creating projectiles and pooling system p9

+++ Assets/Game/Scripts/Core/Configuration/PlayerSettings.cs
@@ -16,7 +16,7 @@ namespace Core.Configuration
 
         public float BulletFireRateInterval;
         public float BulletSpeed;
-        public float BulletDamage;
+        public int BulletDamage;
         public float LaserFireRateInterval;
         public int MaxLaserCharges;
         public float LaserCooldown;
